Fix crash when collecting devDependencies
With the includeDev option set and the following bower.json given:

    {
        "name": "foo",
        "devDependencies": {
            "mocha": "*"
        }
    }

gulp dies with "Cannot set property 'mocha' of undefined".

This is because the current code expects a `dependencies` section to be
provided in bower.json and tries to push `devDependencies` into it. This
patch gives it a default of `{}`.